+++ GroceryMerket.UnitTests/PointOfSaleTerminalShould.cs
@@ -18,7 +18,7 @@ namespace GroceryMarket.UnitTests
         [InlineData(new string[] { "A", "B", "C", "D", "A", "B", "A" }, 13.25)]
         [InlineData(new string[] { "C", "C", "C", "C", "C", "C", "C" }, 6)]
         [InlineData(new string[] { "A", "B", "C", "D" }, 7.25)]
-        public void GetTotalPrices_For_OneOrMoreProducts(string[] products, double expectedTotalPrice)
+        public void GetTotalPrices_For_OneOrMoreProducts(string[] products, decimal expectedTotalPrice)
         {
             // Act
             foreach (string product in products)
